requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
click==8.1.8
coverage==7.6.10
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
lxml==5.3.0
packaging==24.2
peewee==3.17.8
pillow==11.1.0
pluggy==1.5.0
psycopg2-binary==2.9.6
pydantic==2.10.5
pydantic_core==2.27.2
PyPDF2==3.0.1
pytest==8.3.4
pytest-cov==6.0.0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
python-pptx==1.0.2
sniffio==1.3.1
starlette==0.41.3
tomli==2.2.1
typing_extensions==4.12.2
uvicorn==0.34.0
XlsxWriter==3.2.0

transformers~=4.48.0